Authenticator/requirements.txt
Flask==2.2.3
flask_bcrypt==1.0.1
Flask_Cors==3.0.10
flask_restful==0.3.9
flask_sqlalchemy==3.0.3
marshmallow==3.19.0
PyJWT==2.6.0
python-dotenv==1.0.0
Werkzeug==2.2.3
